+++ logger.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/rs/zerolog"
-	"gopkg.in/natefinch/lumberjack.v2"
+	"github.com/shiguredo/lumberjack/v3"
 )
 
 var (