+++ signaling_logger.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/rs/zerolog"
-	"gopkg.in/natefinch/lumberjack.v2"
+	"github.com/shiguredo/lumberjack/v3"
 )
 
 func initSignalingLogger() (*zerolog.Logger, error) {